AWS_Strands/Product_Chandler/requirements.txt
strands-agents>=1.4.0
strands-agents-tools>=0.2.3
python-dotenv>=1.0.0